+++ CHANGELOG.md
@@ -8,10 +8,22 @@
 ### Security
 ### Misc
 
-[Unreleased]: https://github.com/carwow/rspec_logfmt_formatter/compare/v0.0.1...master
+[Unreleased]: https://github.com/carwow/rspec_logfmt_formatter/compare/v0.0.2...master
 
-## [0.0.1] - 2021-01-??
+## [0.0.2] - 2025-07
 
+### Added
+* executable `rspec_logfmt_formatter_collate`
+
+### Changed
+* formatter to `RspecLogfmtFormatter`
+
+### Removed
+* class `RspecLogfmtFormatter::Formatter`
+
+[0.0.2]: https://github.com/carwow/rspec_logfmt_formatter/tree/v0.0.2/
+
+## [0.0.1] - 2021-01
 * Initial release
 
 [0.0.1]: https://github.com/carwow/rspec_logfmt_formatter/tree/v0.0.1/